+++ src/test/java/seedu/meetingjio/MeetingJioTest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package seedu.meetingjio;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class MeetingJioTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
